working on adding testing planets
